requirements.txt
anyio==4.12.1
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hupper==1.12.1
idna==3.11
python-telegram-bot==22.6
flask